reduce compile warnings

+++ inst/include/depth.h
@@ -40,7 +40,7 @@ class width_tree {
   };
 
   phylo_tree_t<node_t> tree;
-  size_t root_no;
+  int root_no;
 
  public:
   explicit width_tree(const std::vector< int >& tree_edge)
@@ -75,7 +75,7 @@ class width_tree {
 
   double calc_b2() {
     double s = 0.0;
-    for (size_t i = 1; i < root_no; ++i) {
+    for (int i = 1; i < root_no; ++i) {
       // we are only interested in tip depths
       s += tree[i].depth / std::pow(2, tree[i].depth);
     }
@@ -85,14 +85,14 @@ class width_tree {
   double var_leaf_depth() {
     double average_depth = 0.0;
     int n = root_no - 1;
-    for (size_t i = 1; i < root_no; ++i) {
+    for (int i = 1; i < root_no; ++i) {
       average_depth += tree[i].depth;
     }
 
     average_depth *= 1.0 / (n);
 
     double var_depth = 0.0;
-    for (size_t i = 1; i < root_no; ++i) {
+    for (int i = 1; i < root_no; ++i) {
       var_depth += (tree[i].depth - average_depth) *
         (tree[i].depth - average_depth);
     }
@@ -102,7 +102,7 @@ class width_tree {
 
   double calc_tot_int_path() {
     double sum_depth = 0.0;
-    for (size_t i = root_no; i < tree.size(); ++i) {
+    for (size_t i = static_cast<size_t>(root_no); i < tree.size(); ++i) {
       sum_depth += tree[i].depth;
     }
     return sum_depth;
